trying to get reviewApp to run componentdidmount before test
